reshuffled autotests
Each test was moved to its own class, and component tests were moved to their separate components.

Some tests had config values added for checking.
