+++ addons/assigngear/XEH_PREP.sqf
@@ -22,6 +22,7 @@ PREP(setInsignia);
 PREP(setUnitTrait);
 PREP(onEdenMessageRecieved);
 PREP(onEdenMissionChange);
+PREP(testGear);
 
 PREP(gui_gearSelector_init);
 PREP(gui_gearSelector_loadFactions);
