+++ addons/assigngear/autotest.hpp
@@ -1,0 +1,6 @@
+class DOUBLES(PREFIX,autotest) {
+    class GVAR(testGear) {
+        code = QUOTE([] call FUNC(testGear));
+        maxWeight = 35; //kg
+    };
+};
