+++ addons/autotest/XEH_PREP.sqf
@@ -1,5 +1,6 @@
 PREP(autotest);
 PREP(testDLC);
 PREP(testEndings);
-PREP(testGear);
 PREP(testInit);
+PREP(testGroups);
+PREP(testAI);
