+++ addons/autotest/functions/fnc_autotest.sqf
@@ -1,58 +1,23 @@
+#define DEBUG_MODE_FULL
 #include "\x\tmf\addons\autotest\script_component.hpp"
 
 private _ctrlListbox = (_this controlsGroupCtrl 101);
 _ctrlListbox lnbSetColumnsPos [0,0.05];
 
-//private _lnbAdd = _ctrlListbox lnbAddRow ["","test"];
-
-//_ctrlListbox lnbSetPicture [[_lnbAdd,0],"\x\tmf\addons\briefing\UI\check_small_ca.paa"];
-
-// Do the tests.
 private _output = [];
-_output append ([] call FUNC(testGear));
-_output append ([] call FUNC(testInit));
-_output append ([] call FUNC(testEndings));
-
-
-//Group count check.
-{
-    private _side = _x;
-    private _groupCount = {(side _x isEqualTo _side)} count allGroups;
-    if (_groupCount > 200) then {
-        _output pushBack [1,format["Side %1 has %2 groups. Note Arma has a 288 group limit per side.",_side,_groupCount]];
-    }
-} forEach [west,east,civilian,resistance];
-
-// AI Numbers check
-
-private _array = playableUnits + [player];
-private _aiCount = {!(_x in _array)} count allUnits;
-
-if (_aiCount > 160) then {
-    if (_aiCount > 200) then {
-        _output pushBack [0,format["You have placed %1 AI. You may wish to consider the performance impact.",_aiCount]];
-    } else {
-        _output pushBack [1,format["You have placed %1 AI. You may wish to consider the performance impact.",_aiCount]];
-    };
-};
-
-// TMF Spectator configured.
-
-if (!((getMissionConfigValue ["respawn",0] == 1) and ("TMF_Spectator" in (getMissionConfigValue ["respawnTemplates",[]])))) then {
-    _output pushBack [0,"TMF Spectator is not setup correctly. See wiki for instructions."];
-};
 
 // Process the Config based auto-tests.
-
 {
     private _test = _x;
+    LOG_1("Running test: %1",configName _x);
     private _code = getText (_test >> "code");
 
     private _outputTest = call compile _code;
     if (_outputTest isEqualType []) then {
         _output append _outputTest;
-    }
-} forEach (configProperties [configFile >> "TMF_autoTest","isClass _x"]);
+        TRACE_2("Appending to output",_output,_outputTest);
+    };
+} forEach ("true" configClasses (configFile >> 'ADDON'));
 
 
 // Display them on the auto-test UI page in Eden.
@@ -67,7 +32,7 @@ if (!((getMissionConfigValue ["respawn",0] == 1) and ("TMF_Spectator" in (getMis
         _ctrlListbox lnbSetPicture [[_lnbAdd,0],QPATHTOEF(briefing,UI\check_small_ca.paa)];
     };
     if (_type == 1) then { //Warning
-        _ctrlListbox lnbSetPicture [[_lnbAdd,0],QPATHTOEF(briefing,UI\warning.paa)];
+        _ctrlListbox lnbSetPicture [[_lnbAdd,0],QPATHTOEF(autotest,UI\warning.paa)];
     };
 
 } forEach _output;
