+++ addons/spectator/XEH_PREP.sqf
@@ -32,3 +32,4 @@ PREP(controlSetPicture);
 PREP(createUnitControl);
 PREP(createVehicleControl);
 PREP(updateKillList);
+PREP(testSpectator);
