+++ addons/spectator/autotest.hpp
@@ -1,0 +1,5 @@
+class DOUBLES(PREFIX,autotest) {
+    class GVAR(testSpectator) {
+        code = QUOTE([] call FUNC(testSpectator));
+    };
+};
